What the functions do
The makeCacheMatrix creates a list of functions relating to the
matrix that it takes as an input and its inverse. The cacheSolve
function solves the inverse and ties it to the list created in
the makeCacheMatrix function. The cacheSolve function also checks
if the inverse has already been solved for and if so it takes it
from the list and does not redo the calculation.